finished circle generator, use radii 25, 65, 85

+++ bboese/bruteBrookBK.cpp
@@ -4,14 +4,14 @@
 using namespace std;
 
 //Prototypes.
-line * calcLine(int p1[], int p2[]);
-void bruteForce(int numPoints, int points[][3]);
+line * calcLine(float p1[], float p2[]);
+void bruteForce(int numPoints, float points[][3]);
 
 int main() {
     int numPoints = 0;
     cin >> numPoints; cin.ignore();
 
-    int points[numPoints][3];
+    float points[numPoints][3];
 
     for (int i = 0; i < numPoints; ++i) {
         cin >> points[i][0]; cin.ignore();
@@ -20,7 +20,7 @@ int main() {
     }
 
 //DISPLAY
-/* 
+/*    
     cout << numPoints << "\n";
     for (int i = 0; i < numPoints; ++i)
         cout << points[i][0] << "\n" << points[i][1] << "\n";
@@ -40,6 +40,7 @@ int main() {
             ++total;
         }
     }
+
     
     cout << total << "\n";
 
@@ -53,13 +54,13 @@ int main() {
     return 0;
 }
 
-line * calcLine(int p1[], int p2[]) {
+line * calcLine(float p1[], float p2[]) {
     line *aLine = NULL;
-    int numer = p1[1] - p2[1];
+    float numer = p1[1] - p2[1];
     if (numer == 0)
         aLine = new horizontal(p1[1]);
     else {
-        int denom = p1[0] - p2[0];
+        float denom = p1[0] - p2[0];
         if (denom == 0)
             aLine = new vertical(p1[0]);
         else {
@@ -69,7 +70,7 @@ line * calcLine(int p1[], int p2[]) {
     return aLine;
 }
 
-void bruteForce(int numPoints, int points[][3]) { 
+void bruteForce(int numPoints, float points[][3]) { 
     for (int i = 0; i < numPoints; ++i) {
         for (int j = i + 1; j < numPoints; ++j) {
             line *aLine = calcLine(points[i], points[j]);
@@ -82,7 +83,7 @@ void bruteForce(int numPoints, int points[][3]) {
                     if (first == 0)
                         first = aLine->whichSide(points[k]);
                     else {
-                        int next = aLine->whichSide(points[k]);
+                        float next = aLine->whichSide(points[k]);
                         if (next != 0 && first != next)
                             SFSGood = false;
                     }
